requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2026.1.4
cffi==2.0.0
curl_cffi==0.14.0
discord-protos==0.0.2
discord.py-self==2.1.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
tzlocal==5.3.1
yarl==1.22.0
